feat: add ifo helper to fix uint8 view issue

+++ packages/lockers/src/common/interfaces/ICakeIFOV8.sol
@@ -12,7 +12,7 @@ interface ICakeIFOV8 {
 
     function addresses(uint256) external view returns (address);
 
-    function computeVestingScheduleIdForAddressAndPid(address holder, uint256 pid) external view returns (bytes32);
+    function computeVestingScheduleIdForAddressAndPid(address holder, uint8 pid) external view returns (bytes32);
 
     function depositPool(uint256 amount, uint8 pid) external;
